Minor FreeBSD build fix
git-svn-id: https://origsvn.digium.com/svn/asterisk/branches/1.4@86787 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ main/astmm.c
@@ -473,7 +473,7 @@ void __ast_mm_init(void)
 		ast_verbose("Asterisk Malloc Debugger Started (see %s))\n", filename);
 	
 	if ((mmlog = fopen(filename, "a+"))) {
-		fprintf(mmlog, "%ld - New session\n", time(NULL));
+		fprintf(mmlog, "%ld - New session\n", (long)time(NULL));
 		fflush(mmlog);
 	}
 }
